Add Server class
Added a server class with a dictionary to map user nicks to their corresponding sockets, and Client class objects.

The selector loop which iterates over sockets that can be read from and/or written to now check if the data associated with the socket is of type Server to determine whether the socket is a server socket. Previously the loop checked if data is of type None to identify the server socket as it had no data associated with. However, now we associated an object of the Server class with the server socket.

MessageParser is not instanciated outside of accept_connection function as it is now solely a wrapper for parseMessage function which returns a typle of containing the parsed message